fix: inspector findings across pkg/tools, pkg/suitability, and backend clients
pkg/tools: consolidate RolePathMiddleware, fix yaml suffix check, CommitTracker int64,
remove dead stubs, unexport readOnlyAllowed, unify I6_VIOLATION error format,
surface git diff failures in ownership check, add isGitModifyCommand and StandardTools tests.

pkg/suitability: add repoRoot existence check, switch default branch, populate
ItemStatus ID/File on error, wire S004 to ClassifyFile read-failure path.

backend clients: use atomic.LoadInt64 in CommitCount() after int64 field change.

+++ pkg/tools/constraints.go
@@ -40,7 +40,7 @@ type Constraints struct {
 }
 
 // CommitTracker tracks git commit invocations for I5 enforcement.
-// The Count field should be accessed atomically (sync/atomic) for goroutine safety.
+// Count must be accessed via sync/atomic (e.g. atomic.AddInt64, atomic.LoadInt64).
 type CommitTracker struct {
-	Count int
+	Count int64
 }
